# src/server/routes/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MeshMonitor API
  description: |
    MeshMonitor REST API v1 provides programmatic access to mesh network data.

    ## Multi-source (breaking change — spec 2.0.0)
    MeshMonitor 4.0 runs one installation against many Meshtastic/MQTT/MeshCore
    sources. Per-resource endpoints are now canonical under
    `/api/v1/sources/{sourceId}/...`:

    | Resource | Canonical path |
    |---|---|
    | Nodes | `/api/v1/sources/{sourceId}/nodes` |
    | Messages | `/api/v1/sources/{sourceId}/messages` |
    | Channels | `/api/v1/sources/{sourceId}/channels` |
    | Telemetry | `/api/v1/sources/{sourceId}/telemetry` |
    | Traceroutes | `/api/v1/sources/{sourceId}/traceroutes` |
    | Packets | `/api/v1/sources/{sourceId}/packets` |
    | Network | `/api/v1/sources/{sourceId}/network` |
    | Position history | `/api/v1/sources/{sourceId}/nodes/{nodeId}/position-history` |
    | Status | `/api/v1/sources/{sourceId}/status` |

    The literal string `default` may be used for `{sourceId}` — it resolves to
    the first source the authenticated token has `read` permission on (or the
    first enabled source for admin tokens). 404 if none is accessible.

    `GET /api/v1/sources` lists the sources the token can read.

    `/api/v1/solar` and `/api/v1/channel-database` remain deployment-global
    (no source scope).

    Legacy root paths (`/api/v1/nodes?sourceId=...`, etc.) were removed in
...
    Use the per-source shape instead.

    ## Authentication
...

    ```
...

...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: |
        API token authentication. Use format: `Bearer mm_v1_<token>`

        Generate tokens from your User Settings page.
...
  schemas:
    Error:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "API token required"
...
    Channel:
      type: object
      properties:
        id:
          type: integer
          description: Channel index (0-7)
          minimum: 0
          maximum: 7
          example: 0
        name:
          type: string
          description: Channel name
          example: "LongFast"
        role:
          type: integer
          description: Channel role (0=Disabled, 1=Primary, 2=Secondary)
          enum: [0, 1, 2]
          example: 1
        roleName:
          type: string
          description: Human-readable role name
          enum: ["Disabled", "Primary", "Secondary"]
          example: "Primary"
        uplinkEnabled:
          type: boolean
          description: Whether MQTT uplink is enabled for this channel
          example: true
        downlinkEnabled:
          type: boolean
          description: Whether MQTT downlink is enabled for this channel
          example: true
        positionPrecision:
          type: integer
          description: Location precision bits (0-32, where 0 means full precision disabled)
          minimum: 0
          maximum: 32
          example: 32
    ChannelListResponse:
      type: object
      required:
        - success
        - count
        - data
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          description: Number of channels returned
          example: 3
        data:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
    ChannelResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Channel'
...
paths:
...
  /channels:
    get:
      tags:
        - Channels
      summary: Get all channels
      description: Retrieve all configured channels in the mesh network
      responses:
        '200':
          description: Channels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /channels/{channelId}:
    get:
      tags:
        - Channels
      summary: Get a specific channel
      description: Retrieve detailed information about a specific channel
      parameters:
        - name: channelId
          in: path
          description: Channel index (0-7)
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 7
      responses:
        '200':
          description: Channel retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...